# packages/contracts/openapi.json
{
  "components": {
    "schemas": {
      "Blocker": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "row_number": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Row Number"
          }
        },
        "required": [
          "row_number",
          "code",
          "message"
        ],
        "title": "Blocker",
        "type": "object"
      },
      "Body_upload_import_imports_post": {
        "properties": {
          "file": {
            "format": "binary",
            "title": "File",
            "type": "string"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_upload_import_imports_post",
        "type": "object"
      },
...
      "ConfirmProblem": {
        "additionalProperties": false,
        "properties": {
          "blockers": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/Blocker"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Blockers"
          },
          "code": {
            "title": "Code",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "portfolio_version_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Portfolio Version Id"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ConfirmProblem",
        "type": "object"
      },
      "ConfirmRequest": {
        "additionalProperties": false,
        "properties": {
          "preview_revision": {
            "minimum": 0.0,
            "title": "Preview Revision",
            "type": "integer"
          }
        },
        "required": [
          "preview_revision"
        ],
        "title": "ConfirmRequest",
        "type": "object"
      },
      "ConfirmResponse": {
        "additionalProperties": false,
        "properties": {
          "excluded_rows": {
            "items": {
              "$ref": "#/components/schemas/ExcludedRow"
            },
            "title": "Excluded Rows",
            "type": "array"
          },
          "import_id": {
            "format": "uuid",
            "title": "Import Id",
            "type": "string"
          },
          "portfolio_version": {
            "$ref": "#/components/schemas/PortfolioVersion"
          },
          "positions": {
            "items": {
              "$ref": "#/components/schemas/Position"
            },
            "title": "Positions",
            "type": "array"
          }
        },
        "required": [
          "portfolio_version",
          "positions",
          "import_id",
          "excluded_rows"
        ],
        "title": "ConfirmResponse",
        "type": "object"
      },
      "Confirmation": {
        "additionalProperties": false,
        "properties": {
          "confirmed_at": {
            "pattern": "^\\d{4}-\\d{2}-\\d{2}T\\d{2}:\\d{2}:\\d{2}(\\.\\d{1,6})?Z$",
            "title": "Confirmed At",
            "type": "string"
          },
          "portfolio_version_id": {
            "format": "uuid",
            "title": "Portfolio Version Id",
            "type": "string"
          }
        },
        "required": [
          "portfolio_version_id",
          "confirmed_at"
        ],
        "title": "Confirmation",
        "type": "object"
      },
...
      "ExcludedRow": {
        "additionalProperties": false,
        "properties": {
          "reason": {
            "enum": [
              "user_excluded",
              "blank_row"
            ],
            "title": "Reason",
            "type": "string"
          },
          "row_number": {
            "title": "Row Number",
            "type": "integer"
          }
        },
        "required": [
          "row_number",
          "reason"
        ],
        "title": "ExcludedRow",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ImportColumn": {
        "additionalProperties": false,
        "properties": {
          "detected": {
            "title": "Detected",
            "type": "string"
          },
          "mapped": {
            "anyOf": [
              {
                "enum": [
                  "symbol",
                  "quantity",
                  "cost_basis",
                  "currency"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mapped"
          },
          "unknown": {
            "title": "Unknown",
            "type": "boolean"
          }
        },
        "required": [
          "detected",
          "mapped",
          "unknown"
        ],
        "title": "ImportColumn",
        "type": "object"
      },
...
      "ImportPreview": {
        "additionalProperties": false,
        "properties": {
          "columns": {
            "items": {
              "$ref": "#/components/schemas/ImportColumn"
            },
            "title": "Columns",
            "type": "array"
          },
          "confirmation": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Confirmation"
              },
              {
                "type": "null"
              }
            ]
          },
          "file_sha256": {
            "title": "File Sha256",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "readiness": {
            "$ref": "#/components/schemas/Readiness"
          },
          "received_at": {
            "format": "date-time",
            "title": "Received At",
            "type": "string"
          },
          "revision": {
            "default": 0,
            "title": "Revision",
            "type": "integer"
          },
          "rows": {
            "items": {
              "$ref": "#/components/schemas/ImportRow"
            },
            "title": "Rows",
            "type": "array"
          },
          "summary": {
            "$ref": "#/components/schemas/ImportSummary"
          }
        },
        "required": [
          "id",
          "file_sha256",
          "received_at",
          "columns",
          "rows",
          "summary"
        ],
        "title": "ImportPreview",
        "type": "object"
      },
      "ImportRow": {
        "additionalProperties": false,
        "properties": {
          "excluded": {
            "default": false,
            "title": "Excluded",
            "type": "boolean"
          },
          "issues": {
            "items": {
              "$ref": "#/components/schemas/ImportIssue"
            },
            "title": "Issues",
            "type": "array"
          },
          "parsed": {
            "$ref": "#/components/schemas/ParsedHolding"
          },
          "raw": {
            "items": {
              "type": "string"
            },
            "title": "Raw",
            "type": "array"
          },
          "resolution": {
            "$ref": "#/components/schemas/Resolution"
          },
          "row_number": {
            "title": "Row Number",
            "type": "integer"
          },
          "status": {
            "enum": [
              "ok",
              "warning",
              "error"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "row_number",
          "raw",
          "parsed",
          "status",
          "issues",
          "resolution"
        ],
        "title": "ImportRow",
        "type": "object"
      },
      "ImportSummary": {
        "additionalProperties": false,
        "properties": {
          "by_code": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "By Code",
            "type": "object"
          },
          "by_resolution": {
            "additionalProperties": {
              "type": "integer"
            },
            "propertyNames": {
              "enum": [
                "resolved",
                "ambiguous",
                "unresolved",
                "not_attempted"
              ]
            },
            "title": "By Resolution",
            "type": "object"
          },
          "by_status": {
            "additionalProperties": {
              "type": "integer"
            },
            "propertyNames": {
              "enum": [
                "ok",
                "warning",
                "error"
              ]
            },
            "title": "By Status",
            "type": "object"
          }
        },
        "required": [
          "by_status",
          "by_code",
          "by_resolution"
        ],
        "title": "ImportSummary",
        "type": "object"
      },
...
      "InstrumentSelection": {
        "additionalProperties": false,
        "properties": {
          "instrument_id": {
            "format": "uuid",
            "title": "Instrument Id",
            "type": "string"
          }
        },
        "required": [
          "instrument_id"
        ],
        "title": "InstrumentSelection",
        "type": "object"
      },
...
      "Problem": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "Problem",
        "type": "object"
      },
      "Readiness": {
        "additionalProperties": false,
        "properties": {
          "blockers": {
            "items": {
              "$ref": "#/components/schemas/Blocker"
            },
            "title": "Blockers",
            "type": "array"
          },
          "confirmable": {
            "title": "Confirmable",
            "type": "boolean"
          }
        },
        "required": [
          "confirmable"
        ],
        "title": "Readiness",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Market Sentinel API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/imports": {
      "post": {
        "description": "Preview a UTF-8 CSV, up to 1 MiB and 1,000 rows.\n\nAccept text/csv or application/csv; also accept application/vnd.ms-excel,\ntext/plain, and application/octet-stream when the filename ends in .csv.\n\nSymbols resolve against synthetic reference data as of the upload's UTC date.\nPreviews are process-local and disappear on restart.",
        "operationId": "upload_import_imports_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_import_imports_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportPreview"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Bad Request"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Request Entity Too Large"
          },
          "415": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Unsupported Media Type"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "summary": "Upload Import",
        "tags": [
          "imports"
        ]
      }
    },
    "/imports/{id}": {
      "get": {
        "operationId": "get_import_imports__id__get",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "title": "Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportPreview"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Import",
        "tags": [
          "imports"
        ]
      }
    },
    "/imports/{id}/confirm": {
      "post": {
        "operationId": "confirm_imports__id__confirm_post",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "title": "Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "idempotency-key",
            "required": true,
            "schema": {
              "maxLength": 200,
              "minLength": 1,
              "pattern": "^[ -~]+$",
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfirmRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfirmResponse"
                }
              }
            },
            "description": "OK"
          },
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfirmResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfirmProblem"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfirmProblem"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfirmProblem"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfirmProblem"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "summary": "Confirm",
        "tags": [
          "imports"
        ]
      }
    },
    "/imports/{id}/rows/{row_number}/exclusion": {
      "delete": {
        "operationId": "include_row_imports__id__rows__row_number__exclusion_delete",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "title": "Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "row_number",
            "required": true,
            "schema": {
              "title": "Row Number",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportRow"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "summary": "Include Row",
        "tags": [
          "imports"
        ]
      },
      "put": {
        "operationId": "exclude_row_imports__id__rows__row_number__exclusion_put",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "title": "Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "row_number",
            "required": true,
            "schema": {
              "title": "Row Number",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportRow"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "summary": "Exclude Row",
        "tags": [
          "imports"
        ]
      }
    },
    "/imports/{id}/rows/{row_number}/resolution": {
      "delete": {
        "description": "Restore automatic resolution on the preview's original date.",
        "operationId": "clear_resolution_imports__id__rows__row_number__resolution_delete",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "title": "Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "row_number",
            "required": true,
            "schema": {
              "title": "Row Number",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportRow"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "summary": "Clear Resolution",
        "tags": [
          "imports"
        ]
      },
      "put": {
        "description": "Select a listed instrument explicitly, retaining automatic issues and candidates.",
        "operationId": "select_resolution_imports__id__rows__row_number__resolution_put",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "title": "Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "row_number",
            "required": true,
            "schema": {
              "title": "Row Number",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InstrumentSelection"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportRow"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "summary": "Select Resolution",
        "tags": [
          "imports"
        ]
      }
    },
...
  }
}